pyproject.toml
[tool.poetry]
name = "swayfade"
version = "0.0.1"
description = "Fades unfocussed windows, for sway and i3"
authors = ["Edd Salkield <edd@salkield.uk>"]

[tool.poetry.dependencies]
python = "^3.7"
i3ipc = "^2.1.1"
toml = "^0.10.0"
pyxdg = "^0.26"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
swayfade = "swayfade.swayfade:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

